Fix value mutator concern

+++ app/Models/Concerns/WithValueMutator.php
@@ -6,6 +6,6 @@ trait WithValueMutator
 {
     public function setValueAttribute(float $value): void
     {
-        $this->attributes['value'] = $value * 100_000_000;
+        $this->attributes['value'] = (int) ($value * 100_000_000);
     }
 }
